fix typing errors & split utils
