+++ pfcli/main.py
@@ -28,7 +28,8 @@ from pfcli.service.client import (
     build_client,
 )
 from pfcli.service.formatter import PanelFormatter
-from pfcli.utils import get_uri, secho_error_and_exit
+from pfcli.utils.format import secho_error_and_exit
+from pfcli.utils.url import get_uri
 
 app = typer.Typer(
     help="Welcome to PeriFlow 🤗",
@@ -160,10 +161,10 @@ def _handle_login_response(r: Response, mfa: bool):
         typer.echo("\n\nLogin success!")
         typer.echo("Welcome back to...")
         typer.echo(" _____          _  _____ _")
-        typer.echo("|  __ \___ _ __(_)|  ___| | _____      __")
-        typer.echo("|  ___/ _ \ '__| || |__ | |/ _ \ \ /\ / /")
+        typer.echo("|  __ \___ _ __(_)|  ___| | _____      __") # type: ignore
+        typer.echo("|  ___/ _ \ '__| || |__ | |/ _ \ \ /\ / /") # type: ignore
         typer.echo("| |  |  __/ |  | ||  __|| | (_) | V  V / ")
-        typer.echo("|_|   \___|_|  |_||_|   |_|\___/ \_/\_/  ")
+        typer.echo("|_|   \___|_|  |_||_|   |_|\___/ \_/\_/  ") # type: ignore
         typer.echo("\n\n")
     except HTTPError:
         if mfa:
